spin out ArticleHeadlinePadding into own file
