+++ src/web/components/ArticleHeadline.tsx
@@ -60,14 +60,6 @@ const lightFont = css`
     }
 `;
 
-const standardPadding = css`
-    padding-bottom: 24px;
-    padding-top: 3px;
-    ${from.tablet} {
-        padding-bottom: 36px;
-    }
-`;
-
 const underlinedStyles = css`
     background-image: repeating-linear-gradient(
         to bottom,
@@ -251,37 +243,6 @@ const renderHeadline = ({
                 </h1>
             );
     }
-};
-
-const determinPadding = (designType: DesignType) => {
-    switch (designType) {
-        case 'Article':
-        case 'Media':
-        case 'Live':
-        case 'SpecialReport':
-        case 'Recipe':
-        case 'MatchReport':
-        case 'GuardianView':
-        case 'GuardianLabs':
-        case 'Quiz':
-        case 'AdvertisementFeature':
-        case 'Review':
-        case 'Feature':
-        case 'Comment':
-        case 'Analysis':
-            return standardPadding;
-        case 'Interview':
-        case 'Immersive':
-            return null;
-    }
-};
-
-export const ArticleHeadlineWrapper: React.FC<{
-    children: React.ReactNode;
-    designType: DesignType;
-}> = ({ children, designType }) => {
-    const paddingClassName = determinPadding(designType);
-    return <div className={paddingClassName || ''}>{children}</div>;
 };
 
 export const ArticleHeadline = ({
